backend/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
attrs==20.3.0
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
constantly==15.1.0
cryptography==3.3.1
cssselect==1.1.0
distlib==0.3.1
distro==1.5.0
Django==3.1.5
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.2
environ==1.0
et-xmlfile==1.0.1
filelock==3.0.12
hyperlink==21.0.0
idna==2.10
importlib-metadata==3.4.0
incremental==17.5.0
install==1.3.4
itemadapter==0.2.0
itemloaders==1.0.4
jdcal==1.4.1
jmespath==0.10.0
lxml==4.6.2
Markdown==3.3.3
mysql-server-has-gone-away==2.0.0
mysqlclient==2.0.3
numpy==1.19.5
openpyxl==3.0.6
pandas==1.2.1
parsel==1.6.0
Pillow==8.1.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
pytesseract==0.3.7
python-dateutil==2.8.1
python-http-client==3.3.1
pytz==2020.5
queuelib==1.5.0
requests==2.25.1
Scrapy==2.4.1
selenium==3.141.0
sendgrid==6.5.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
tabula-py==2.2.0
tesseract==0.1.3
Twisted==20.3.0
typing-extensions==3.7.4.3
urllib3==1.26.2
virtualenv==20.3.1
w3lib==1.22.0
wrapt==1.12.1
zipp==3.4.0
zope.interface==5.2.0
